@@ -2,7 +2,6 @@ package hope
 
 import (
 	"errors"
-	"io/ioutil"
 	"os"
 	"path"
 )
@@ -41,7 +40,7 @@ func FetchKubeconfig(log *logrus.Entry, node *Node, merge bool) error {
 		}
 
 		if !merge {
-			return errors.New("Refusing to overwrite existing kubeconfig file.")
+			return errors.New("refusing to overwrite existing kubeconfig file")
 		}
 	} else if os.IsNotExist(err) {
 		if err := os.MkdirAll(path.Dir(kubeconfigFile), 0700); err != nil {
@@ -66,7 +65,7 @@ func FetchKubeconfig(log *logrus.Entry, node *Node, merge bool) error {
 		return err
 	}
 
-	err = ioutil.WriteFile(kubeconfigFile, []byte(kubeconfigContents), 0600)
+	err = os.WriteFile(kubeconfigFile, []byte(kubeconfigContents), 0600)
 	if err != nil {
 		return err
 	}
