@@ -11,9 +11,7 @@ var UseSudo bool = false
 
 // TODO: Add a bunch of methods that wrap individual commands.
 // Invocations require the package name anyways, so something like:
-//   docker.Pull(...)
-// is a lot nicer than
-//   docker.ExecDocker("pull", ...)
+// docker.Pull(...) is a lot nicer than docker.ExecDocker("pull", ...)
 type ExecDockerFunc func(args ...string) error
 type GetDockerFunc func(args ...string) (string, error)
 
